# src/flask_app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sistem Planificare Examene API (Flask)",
    "description": "API pentru sistemul de planificare examene FIESC implementat în Flask",
    "version": "1.0.0",
    "contact": {
      "email": "planificare@fiesc.usv.ro"
    }
  },
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Autentificare cu Google OAuth",
        "description": "Endpoint pentru autentificare cu Google OAuth",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Token Google OAuth",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "token"
              ],
              "properties": {
                "token": {
                  "type": "string",
                  "example": "google_oauth_token"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autentificare reușită",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string",
                  "example": "jwt_token"
                },
                "token_type": {
                  "type": "string",
                  "example": "bearer"
                },
                "expires_in": {
                  "type": "integer",
                  "example": 3600
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 123
                    },
                    "email": {
                      "type": "string",
                      "example": "user@usv.ro"
                    },
                    "role": {
                      "type": "string",
                      "example": "SEC"
                    },
                    "firstName": {
                      "type": "string",
                      "example": "Nume"
                    },
                    "lastName": {
                      "type": "string",
                      "example": "Prenume"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Token Google OAuth lipsă"
          },
          "401": {
            "description": "Token Google OAuth invalid sau email invalid pentru aplicație"
          },
          "500": {
            "description": "Eroare la crearea utilizatorului"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Obținerea utilizatorului curent",
        "description": "Endpoint pentru obținerea utilizatorului curent",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Utilizator obținut cu succes",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "example": 123
                },
                "email": {
                  "type": "string",
                  "example": "user@usv.ro"
                },
                "role": {
                  "type": "string",
                  "example": "SEC"
                },
                "firstName": {
                  "type": "string",
                  "example": "Nume"
                },
                "lastName": {
                  "type": "string",
                  "example": "Prenume"
                }
              }
            }
          },
          "401": {
            "description": "Utilizator neautentificat"
          },
          "404": {
            "description": "Utilizator negăsit"
          },
          "500": {
            "description": "Eroare la obținerea utilizatorului"
          }
        }
      }
    },
...
  },
...
}